Update minesweeper.py
Introduced a helper function getValidInt() that prompts the user for an integer input until they provide a valid one. This function also allows specifying minimum and maximum values.

Modified the main() function to use the new helper function for getting the board size and number of mines. For the number of mines, we now also ensure that the user cannot specify more mines than there are cells on the board.